+++ src/plugins/cdrtools/burn-mkisofs.c
@@ -61,7 +61,7 @@ static GObjectClass *parent_class = NULL;
 static BraseroBurnResult
 brasero_mkisofs_read_isosize (BraseroProcess *process, const gchar *line)
 {
-	gint sectors;
+	gint64 sectors;
 
 	sectors = strtoll (line, NULL, 10);
 	if (!sectors)
@@ -70,7 +70,7 @@ brasero_mkisofs_read_isosize (BraseroProcess *process, const gchar *line)
 	/* mkisofs reports blocks of 2048 bytes */
 	brasero_job_set_output_size_for_current_track (BRASERO_JOB (process),
 						       sectors,
-						       sectors * 2048);
+						       sectors * 2048ULL);
 	return BRASERO_BURN_OK;
 }
